+++ RayTracing/Object.h
@@ -2,12 +2,12 @@
 #include "Vector.h"
 #include "Ray.h"
 struct Hit_record {
-	float t;
+	double t;
 	Vector3f p;
 	Vector3f normal;
 };
 
 class Object {
 public:
-	virtual bool intersect(const Ray&, float tmin, float tmax, Hit_record&) const = 0;
+	virtual bool intersect(const Ray&, double tmin, double tmax, Hit_record&) const = 0;
 };
