+++ RayTracing/ObjectList.h
@@ -6,5 +6,5 @@ class ObjectList {
 	int list_size;
 public:
 	ObjectList(Object** l, int n) : list(l), list_size(n){}
-	bool intersect(const Ray&, float tmin, float tmax, Hit_record&) const;
+	bool intersect(const Ray&, double tmin, double tmax, Hit_record&) const;
 };
