+++ RayTracing/Sphere.h
@@ -2,8 +2,8 @@
 #include "Object.h"
 class Sphere : public Object {
 	Vector3f center;
-	float radius;
+	double radius;
 public:
-	bool intersect(const Ray&, float tmin, float tmax, Hit_record&) const override;
-	Sphere(const Vector3f& v = Vector3f(0,0,0), float r = 1.f) : center(v), radius(r){}
+	bool intersect(const Ray&, double tmin, double tmax, Hit_record&) const override;
+	Sphere(const Vector3f& v = Vector3f(0,0,0), double r = 1.) : center(v), radius(r){}
 };
